football/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.8
djangorestframework==3.12.4
whitenoise==5.3.0
idna==3.2
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
gunicorn==20.1.0
django-cors-headers==3.10.0
